# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/app/goods/detail": {
            "get": {
                "description": "传入 id",
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "获取商品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "goods_id",
                        "name": "goods_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/app.GoodsDetail"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/goods/hot": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "获取热门商品列表",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "arrary"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/goods/list": {
            "get": {
                "description": "传入 category_id | kewords | name",
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "获取商品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page_num",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page_size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/app.GoodsList"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/goods/search": {
            "get": {
                "description": "传入   name | kewords",
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "搜索商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page_num",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page_size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/app.GoodsList"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/app/order/list": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取订单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订单状态",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_num",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.Order"
                            }
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/order/submit": {
            "post": {
                "description": "从购物车中提交订单",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "提交订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "messge 备注",
                        "name": "messge",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "选择地址的id",
                        "name": "address_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/app/order/update": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改订单状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "open_id",
                        "name": "open_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订单号",
                        "name": "order_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订单状态",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "app.GoodsDetail": {
            "type": "object",
            "properties": {
                "brief": {
                    "description": "商品简介",
                    "type": "string"
                },
                "category_id": {
                    "description": "商品所属类目ID",
                    "type": "integer"
                },
                "detail": {
                    "description": "商品详细介绍，是富文本格式",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keywords": {
                    "description": "商品关键字，采用逗号间隔",
                    "type": "string"
                },
                "name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "options": {
                    "description": "商品购买选项",
                    "type": "string"
                },
                "pic_url": {
                    "description": "商品主图",
                    "type": "string"
                },
                "pic_urls": {
                    "description": "商品附图JSON数组格式",
                    "type": "string"
                },
                "price": {
                    "description": "价格",
                    "type": "number"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "unit": {
                    "description": "商品单位，例如件、盒",
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "app.GoodsList": {
            "type": "object",
            "properties": {
                "brief": {
                    "description": "商品简介",
                    "type": "string"
                },
                "category_id": {
                    "description": "商品所属类目ID",
                    "type": "integer"
                },
                "counter_price": {
                    "description": "专柜价格",
                    "type": "number"
                },
                "detail": {
                    "description": "商品详细介绍，是富文本格式",
                    "type": "string"
                },
                "gallery": {
                    "description": "商品宣传图片列表，采用JSON数组格式",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_hot": {
                    "description": "是否人气推荐，如果设置则可以在人气推荐页面展示",
                    "type": "integer"
                },
                "is_on_sale": {
                    "description": "是否上架",
                    "type": "integer"
                },
                "keywords": {
                    "description": "商品关键字，采用逗号间隔",
                    "type": "string"
                },
                "name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "options": {
                    "description": "商品购买选项",
                    "type": "string"
                },
                "pic_url": {
                    "description": "商品页面商品图片",
                    "type": "string"
                },
                "price": {
                    "description": "线上价格",
                    "type": "number"
                },
                "retail_price": {
                    "description": "零售价格",
                    "type": "number"
                },
                "unit": {
                    "description": "商品单位，例如件、盒",
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "app.Order": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "description": "实付费用， = order_price - integral_price",
                    "type": "number"
                },
                "address": {
                    "description": "收货具体地址",
                    "type": "string"
                },
                "confirm_time": {
                    "description": "用户确认收货时间",
                    "type": "string"
                },
                "coupon_price": {
                    "description": "优惠券减免",
                    "type": "number"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deleted": {
                    "description": "逻辑删除",
                    "type": "integer"
                },
                "deleted_time": {
                    "description": "删除时间",
                    "type": "string"
                },
                "dis_price": {
                    "description": "配送费用",
                    "type": "number"
                },
                "end_time": {
                    "description": "订单关闭时间",
                    "type": "string"
                },
                "goods_count": {
                    "description": "商品总数量",
                    "type": "integer"
                },
                "goods_info": {
                    "description": "商品信息json",
                    "type": "string"
                },
                "goods_price": {
                    "description": "商品总费用",
                    "type": "number"
                },
                "groupon_price": {
                    "description": "团购优惠价减免",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "integral_price": {
                    "description": "用户积分减免",
                    "type": "number"
                },
                "message": {
                    "description": "用户订单留言",
                    "type": "string"
                },
                "open_id": {
                    "description": "用户表的用户ID",
                    "type": "string"
                },
                "order_price": {
                    "description": "订单费用， = goods_price + dis_price - coupon_price",
                    "type": "number"
                },
                "pay_id": {
                    "description": "微信付款编号",
                    "type": "string"
                },
                "pay_time": {
                    "description": "微信付款时间",
                    "type": "string"
                },
                "ship_channel": {
                    "description": "外卖平台",
                    "type": "integer"
                },
                "ship_sn": {
                    "description": "外卖订单",
                    "type": "string"
                },
                "ship_time": {
                    "description": "发货开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "订单状态 1已提交 2已完成 3撤销 4已接单不可撤销",
                    "type": "integer"
                },
                "sub_status": {
                    "description": "订单子状态 1已提交 2已完成 3撤销 4已接单不可撤销",
                    "type": "integer"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}